pyproject.toml
[tool.poetry]
name = "databricks-langchain-agent"
version = "0.1.0"
description = "Databricks LangChain agent with tool calling capabilities"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
keywords = ["databricks", "langchain", "langgraph", "mlflow", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
databricks-agents = "^1.8.2"
databricks-langchain = "^0.11.0"
langchain = "^1.0.0"
langchain-core = "^1.0.0"
langgraph = "^1.0.3"
mlflow-skinny = {extras = ["databricks"], version = "^3.6.0"}
python-dotenv = "^1.2.1"
langchain-openai = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
